Add error handling to boost::filesystem usage

+++ indra/llfilesystem/lldiskcache.cpp
@@ -80,16 +80,28 @@ void LLDiskCache::purge()
 #else
     std::string cache_path(sCacheDir);
 #endif
-    if (boost::filesystem::is_directory(cache_path))
+    boost::system::error_code ec;
+    if (boost::filesystem::is_directory(cache_path, ec) && !ec.failed())
     {
         for (auto& entry : boost::make_iterator_range(boost::filesystem::recursive_directory_iterator(cache_path), {}))
         {
-            if (boost::filesystem::is_regular_file(entry))
+            ec.clear();
+            if (boost::filesystem::is_regular_file(entry, ec) && !ec.failed())
             {
                 if (entry.path().string().rfind(sCacheFilenameExt) != std::string::npos)
                 {
-                    uintmax_t file_size = boost::filesystem::file_size(entry);
-                    const std::time_t file_time = boost::filesystem::last_write_time(entry);
+                    const uintmax_t file_size = boost::filesystem::file_size(entry, ec);
+                    if (ec.failed())
+                    {
+                        LL_WARNS() << "Failed to read file size for cache file " << entry.path().string() << ": " << ec.message() << LL_ENDL;
+                        continue;
+                    }
+                    const std::time_t file_time = boost::filesystem::last_write_time(entry, ec);
+					if (ec.failed())
+					{
+						LL_WARNS() << "Failed to read last write time for cache file " << entry.path().string() << ": " << ec.message() << LL_ENDL;
+                        continue;
+					}
 
                     file_info.push_back(file_info_t(file_time, { file_size, entry.path() }));
                 }
@@ -118,6 +130,7 @@ void LLDiskCache::purge()
             if (ec.failed())
             {
                 LL_WARNS() << "Failed to delete cached file " << entry.second.second << ": " << ec.message() << LL_ENDL;
+                continue;
             }
         }
 
@@ -223,13 +236,13 @@ void LLDiskCache::clearCache()
 #else
     boost::filesystem::path cache_path(sCacheDir);
 #endif
-    if (boost::filesystem::is_directory(cache_path))
+    boost::system::error_code ec;
+    if (boost::filesystem::is_directory(cache_path, ec) && !ec.failed())
     {
-        boost::system::error_code ec;
         boost::filesystem::remove_all(cache_path, ec);
         if (ec.failed())
         {
-            LL_WARNS() << "Failed to delete cached files " << cache_path << ": " << ec.message() << LL_ENDL;
+            LL_WARNS() << "Failed to delete cached files " << cache_path << " : " << ec.message() << LL_ENDL;
         }
 
         createCache();
@@ -254,15 +267,23 @@ uintmax_t LLDiskCache::dirFileSize(const std::string dir)
 #else
     boost::filesystem::path dir_path(dir);
 #endif
-    if (boost::filesystem::is_directory(dir_path))
+    boost::system::error_code ec;
+    if (boost::filesystem::is_directory(dir_path, ec) && !ec.failed())
     {
         for (auto& entry : boost::make_iterator_range(boost::filesystem::recursive_directory_iterator(dir_path), {}))
         {
-            if (boost::filesystem::is_regular_file(entry))
+            ec.clear();
+            if (boost::filesystem::is_regular_file(entry, ec) && !ec.failed())
             {
                 if (entry.path().string().rfind(sCacheFilenameExt) != std::string::npos)
                 {
-                    total_file_size += boost::filesystem::file_size(entry);
+                    uintmax_t file_size = boost::filesystem::file_size(entry, ec);
+                    if (ec.failed())
+                    {
+                        LL_WARNS() << "Failed to get file size for cache file " << entry.path().string() << " : " << ec.message() << LL_ENDL;
+                        continue;
+                    }
+                    total_file_size += file_size;
                 }
             }
         }
